Add partial service method template
- Introduce new partial template for generating public and private
  service methods
- Introduce a service method generator
- Consolidate template logic into reusable execute function
