+++ internal/service_method_generator.go
@@ -1,0 +1,17 @@
+package internal
+
+import (
+	"io"
+)
+
+type ServiceMethodGenerator struct {
+	MethodName  string
+	InputName   string
+	OutputName  string
+	PackageName string
+	Private     bool
+}
+
+func (g *ServiceMethodGenerator) Generate(w io.Writer) error {
+	return execute("service_method", templateServiceMethod, w, g)
+}
