+++ internal/service_register_generator.go
@@ -1,9 +1,7 @@
 package internal
 
 import (
-	"fmt"
 	"io"
-	"text/template"
 
 	"google.golang.org/protobuf/compiler/protogen"
 )
@@ -16,26 +14,5 @@ type ServiceRegisterGenerator struct {
 }
 
 func (g *ServiceRegisterGenerator) Generate(w io.Writer) error {
-	tmpl := template.Must(
-		template.
-			New("service_register").
-			Funcs(templateFuncs()).
-			Parse(templateServiceRegister),
-	)
-
-	return tmpl.Execute(w, g)
-}
-
-func templateFuncs() template.FuncMap {
-	return template.FuncMap{
-		"sprintf": fmt.Sprintf,
-		"previous": func(current *Service, services []*Service) *Service {
-			for i, service := range services {
-				if current == service {
-					return services[i-1]
-				}
-			}
-			return nil
-		},
-	}
+	return execute("service_register", templateServiceRegister, w, g)
 }
